+++ backend/event/mutations/CreateEpisodeEntityLinkMutation.py
@@ -14,7 +14,7 @@ from django.core.exceptions import ValidationError
 from django.db import transaction
 from core.types.entity import Entity
 from core.entity_models import ENTITY_MODELS
-from source.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
+from user.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
 
 
 class CreateEpisodeEntityLinkInput(InputObjectType):
