+++ backend/event/mutations/CreateEpisodeMutation.py
@@ -13,7 +13,7 @@ from event.types.EpisodeType import EpisodeType
 from graphql_app.LettercraftMutation import LettercraftMutation
 from graphql_app.types.LettercraftErrorType import LettercraftErrorType
 from source.models import Source
-from source.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
+from user.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
 
 
 class CreateEpisodeInput(InputObjectType):
