+++ backend/event/mutations/DeleteEpisodeMutation.py
@@ -3,7 +3,7 @@ from graphene import ID, Boolean, List, Mutation, NonNull, ResolveInfo
 from event.models import Episode
 from event.types.EpisodeType import EpisodeType
 from graphql_app.types.LettercraftErrorType import LettercraftErrorType
-from source.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
+from user.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
 
 class DeleteEpisodeMutation(Mutation):
     ok = Boolean(required=True)
