+++ backend/event/mutations/UpdateEpisodeMutation.py
@@ -5,7 +5,7 @@ from core.types.input.EntityDescriptionInputType import EntityDescriptionInputTy
 from event.models import Episode
 from graphql_app.LettercraftMutation import LettercraftMutation
 from graphql_app.types.LettercraftErrorType import LettercraftErrorType
-from source.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
+from user.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
 
 
 class UpdateEpisodeInput(EntityDescriptionInputType, InputObjectType):
