+++ backend/letter/mutations/UpdateGiftMutation.py
@@ -4,7 +4,7 @@ from core.types.input.DescriptionFieldInputType import DescriptionFieldInputType
 from core.types.input.EntityDescriptionInputType import EntityDescriptionInputType
 from letter.models import GiftDescription
 from graphql_app.LettercraftMutation import LettercraftMutation
-from source.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
+from user.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
 from graphql_app.types.LettercraftErrorType import LettercraftErrorType
