+++ backend/space/mutations/UpdateSpaceMutation.py
@@ -5,7 +5,7 @@ from graphql_app.LettercraftMutation import LettercraftMutation
 
 from graphql_app.types.LettercraftErrorType import LettercraftErrorType
 from space.models import SpaceDescription
-from source.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
+from user.permissions import can_edit_source, SOURCE_NOT_PERMITTED_MSG
 
 
 class UpdateSpaceInput(EntityDescriptionInputType, InputObjectType):
